[VL-1] Complete DAO tests

+++ video-library-adapter/jdbc-adapter/src/main/java/it/proconsole/library/video/adapter/jdbc/model/FilmReview.java
@@ -4,5 +4,5 @@ import org.springframework.lang.Nullable;
 
 import java.time.LocalDateTime;
 
-public record FilmReview(long id, LocalDateTime date, int rating, @Nullable String detail, int filmId) {
+public record FilmReview(Long id, LocalDateTime date, int rating, @Nullable String detail, Long filmId) {
 }
