+++ video-library-adapter/jdbc-adapter/src/main/java/it/proconsole/library/video/adapter/jdbc/repository/entity/FilmReviewEntity.java
@@ -5,9 +5,14 @@ import org.springframework.lang.Nullable;
 
 import java.time.LocalDateTime;
 
-public record FilmReviewEntity(Long id, LocalDateTime date, Integer rating, @Nullable String detail,
-                               Integer filmId) implements EntityWithId {
-  public FilmReviewEntity(LocalDateTime date, int rating, @Nullable String detail, int filmId) {
+public record FilmReviewEntity(
+    Long id,
+    LocalDateTime date,
+    Integer rating,
+    @Nullable String detail,
+    Long filmId
+) implements EntityWithId {
+  public FilmReviewEntity(LocalDateTime date, int rating, @Nullable String detail, Long filmId) {
     this(1L, date, rating, detail, filmId);
   }
